+++ ember_mug/data.py
@@ -2,9 +2,20 @@
 from __future__ import annotations
 
 from dataclasses import asdict, dataclass, is_dataclass
+from functools import cached_property
 from typing import Any, NamedTuple
 
-from .consts import ATTR_LABELS, EXTRA_ATTRS, LiquidState, TemperatureUnit
+from .consts import (
+    ATTR_LABELS,
+    EMBER_CUP,
+    EMBER_TRAVEL_MUG,
+    EXTRA_ATTRS,
+    INITIAL_ATTRS,
+    TRAVEL_MUG_ATTRS,
+    UPDATE_ATTRS,
+    LiquidState,
+    TemperatureUnit,
+)
 from .formatting import format_led_colour, format_liquid_level, format_temp
 from .utils import bytes_to_little_int, decode_byte_string
 
@@ -96,6 +107,49 @@ class MugFirmwareInfo:
 
 
 @dataclass
+class Model:
+    """Model name and attributes based on mode."""
+
+    name: str
+    include_extra: bool = False
+
+    @cached_property
+    def is_cup(self) -> bool:
+        """Check if the model is a Cup."""
+        return self.name.startswith(EMBER_CUP)
+
+    @cached_property
+    def is_travel_mug(self) -> bool:
+        """Check if the model is a Travel mug."""
+        return self.name.startswith(EMBER_TRAVEL_MUG)
+
+    @cached_property
+    def attribute_labels(self) -> dict[str, str]:
+        """Calculated labels for includes attributes."""
+        all_attrs = self.initial_attributes | self.update_attributes | {'use_metric'}
+        return {attr: label for attr, label in ATTR_LABELS.items() if attr in all_attrs}
+
+    @cached_property
+    def initial_attributes(self) -> set[str]:
+        """Initial attributes based on model and extra."""
+        if self.include_extra is False:
+            return INITIAL_ATTRS - EXTRA_ATTRS
+        return INITIAL_ATTRS
+
+    @cached_property
+    def update_attributes(self) -> set[str]:
+        """Attributes to update based on model and extra."""
+        attributes = UPDATE_ATTRS
+        if self.include_extra is False:
+            attributes = attributes - EXTRA_ATTRS
+        if self.is_cup:
+            attributes = attributes - {'name'}
+        elif self.is_travel_mug:
+            attributes = attributes | TRAVEL_MUG_ATTRS
+        return attributes
+
+
+@dataclass
 class MugMeta:
     """Meta data for mug."""
 
@@ -120,9 +174,8 @@ class MugData:
     """Class to store/display the state of the mug."""
 
     # Options
-    model: str
+    model: Model
     use_metric: bool = True
-    include_extra: bool = False
 
     # Attributes
     name: str = ""
@@ -143,7 +196,7 @@ class MugData:
     @property
     def meta_display(self) -> str:
         """Return Meta infor based on preference."""
-        if self.meta and not self.include_extra:
+        if self.meta and not self.model.include_extra:
             return f'Serial Number: {self.meta.serial_number}'
         return str(self.meta)
 
@@ -190,11 +243,7 @@ class MugData:
     @property
     def formatted(self) -> dict[str, Any]:
         """Return human-readable names and values for all attributes for display."""
-        return {
-            label: self.get_formatted_attr(attr)
-            for attr, label in ATTR_LABELS.items()
-            if self.include_extra or attr not in EXTRA_ATTRS
-        }
+        return {label: self.get_formatted_attr(attr) for attr, label in self.model.attribute_labels.items()}
 
     def as_dict(self) -> dict[str, Any]:
         """Dump all attributes as dict for info/debugging."""