+++ tests/conftest.py
@@ -8,10 +8,11 @@ import pytest_asyncio
 from bleak.backends.device import BLEDevice
 
 from ember_mug import EmberMug
-from ember_mug.data import MugData
+from ember_mug.consts import EMBER_MUG
+from ember_mug.data import Model, MugData
 
 TEST_MAC = '32:36:a5:be:88:cb'
-TEST_MODEL_NAME = 'Ember Ceramic Mug'
+TEST_MODEL_NAME = EMBER_MUG
 
 
 class AsyncContextManager:
@@ -36,7 +37,7 @@ def ble_device_fixture() -> Generator[BLEDevice, None, None]:
 
 @pytest.fixture
 def mug_data(ble_device: BLEDevice) -> Generator[MugData, None, None]:
-    yield MugData(ble_device.name or TEST_MODEL_NAME)
+    yield MugData(Model(ble_device.name or TEST_MODEL_NAME))
 
 
 @pytest_asyncio.fixture