+++ tests/test_data.py
@@ -1,7 +1,8 @@
 """Tests for `ember_mug.data`."""
 from __future__ import annotations
 
-from ember_mug.data import BatteryInfo, Change, Colour, MugFirmwareInfo, MugMeta
+from ember_mug.consts import EMBER_CUP, EMBER_MUG, EMBER_TRAVEL_MUG
+from ember_mug.data import BatteryInfo, Change, Colour, Model, MugFirmwareInfo, MugMeta
 
 
 def test_change() -> None:
@@ -37,3 +38,29 @@ def test_mug_meta() -> None:
     assert meta.mug_id == 'WXc9PT09'
     assert meta.serial_number == 'ABCDEFGHIJ'
     assert str(meta) == 'Mug ID: WXc9PT09, Serial Number: ABCDEFGHIJ'
+
+
+def test_mug_model() -> None:
+    mug = Model(EMBER_MUG)
+    assert mug.is_travel_mug is False
+    assert mug.is_travel_mug is False
+    assert 'udsk' not in mug.initial_attributes
+    assert 'name' in mug.update_attributes
+    assert 'name' in mug.attribute_labels
+    assert 'battery_voltage' not in mug.update_attributes
+    mug_with_extra = Model(EMBER_MUG, include_extra=True)
+    assert 'udsk' in mug_with_extra.initial_attributes
+
+    travel_mug = Model(EMBER_TRAVEL_MUG)
+    assert travel_mug.is_travel_mug is True
+    assert travel_mug.is_cup is False
+    assert 'name' in travel_mug.update_attributes
+    assert 'name' in travel_mug.attribute_labels
+    assert 'battery_voltage' in travel_mug.attribute_labels
+
+    cup = Model(EMBER_CUP)
+    assert cup.is_cup is True
+    assert cup.is_travel_mug is False
+    assert 'name' not in cup.update_attributes
+    assert 'name' not in cup.attribute_labels
+    assert 'battery_voltage' not in cup.update_attributes